+++ zcash_client_backend/src/data_api/chain.rs
@@ -147,17 +147,16 @@
 use std::ops::Range;
 
 use zcash_primitives::{
-    block::BlockHash,
     consensus::{self, BlockHeight},
     sapling::{self, note_encryption::PreparedIncomingViewingKey},
     zip32::Scope,
 };
 
 use crate::{
-    data_api::{BlockMetadata, NullifierQuery, WalletWrite},
+    data_api::{NullifierQuery, WalletWrite},
     proto::compact_formats::CompactBlock,
     scan::BatchRunner,
-    scanning::{add_block_to_runner, check_continuity, scan_block_with_runner, ScanningKey},
+    scanning::{add_block_to_runner, scan_block_with_runner, ScanningKey},
 };
 
 pub mod error;
@@ -326,46 +325,10 @@ where
         None
     };
 
-    let mut continuity_check_metadata = prior_block_metadata;
     block_source.with_blocks::<_, DbT::Error>(
         Some(from_height),
         Some(limit),
         |block: CompactBlock| {
-            // check block continuity
-            if let Some(scan_error) = check_continuity(&block, continuity_check_metadata.as_ref()) {
-                return Err(Error::Scan(scan_error));
-            }
-
-            if block.height() == BlockHeight::from(0) {
-                // We can always derive a valid `continuity_check_metadata` for the
-                // genesis block, even if the block source doesn't have
-                // `sapling_commitment_tree_size`. So briefly set it to a dummy value that
-                // ensures the `map` below produces the correct genesis block value.
-                assert!(continuity_check_metadata.is_none());
-                continuity_check_metadata = Some(BlockMetadata::from_parts(
-                    BlockHeight::from(0),
-                    BlockHash([0; 32]),
-                    0,
-                ));
-            }
-            continuity_check_metadata = continuity_check_metadata.as_ref().map(|m| {
-                BlockMetadata::from_parts(
-                    block.height(),
-                    block.hash(),
-                    block
-                        .chain_metadata
-                        .as_ref()
-                        .map(|m| m.sapling_commitment_tree_size)
-                        .unwrap_or_else(|| {
-                            m.sapling_tree_size()
-                                + u32::try_from(
-                                    block.vtx.iter().map(|tx| tx.outputs.len()).sum::<usize>(),
-                                )
-                                .unwrap()
-                        }),
-                )
-            });
-
             add_block_to_runner(params, block, &mut batch_runner);
 
             Ok(())
@@ -415,7 +378,7 @@ where
                     .map(|out| (out.account(), *out.nf()))
             }));
 
-            prior_block_metadata = Some(*scanned_block.metadata());
+            prior_block_metadata = Some(scanned_block.to_block_metadata());
             scanned_blocks.push(scanned_block);
 
             Ok(())